debugging

+++ tools/write_dataset.py
@@ -32,8 +32,8 @@ if __name__ == "__main__":
     writer = DatasetWriter(write_path, {
         # Tune options to optimize dataset size, throughput at train-time
         "images": NDArrayField(dtype=np.dtype("int32"), shape=(3, 3, 227, 227)),
-        "domains": NDArrayField(dtype=np.dtype("int32"), shape=(3, len(domains))),
-        "contents": NDArrayField(dtype=np.dtype("int32"), shape=(3, len(contents)))
+        "domains": NDArrayField(dtype=np.dtype("int64"), shape=(3, len(domains))),
+        "contents": NDArrayField(dtype=np.dtype("int64"), shape=(3, len(contents)))
     }, num_workers=20)
 
     # Write dataset
